# contracts/openapi/scene-service.yaml
...
openapi: 3.0.3
info:
  title: HunterEdge - scene-service
  version: 0.1.0
  description: |
    场景生成服务统一契约（端口 8081，网关前缀 `/api/v1/scene`）。

    **统一响应**：`{code, message, data, request_id, timestamp}`；`code=0` 成功，非 0 为预定义错误码（附录 A）。

    **场景分类**（4.2.1 节）：基础场景（直线巡航 / 弯道行驶 / 跟车行驶 / 变道行驶）、
    交互场景（前车切入 / 行人横穿 / 路口会车 / 施工绕行）、环境场景（雨天 / 夜间 / 雾天 / 强光逆光）、
    极端场景 Corner Case（紧急制动 / 障碍物突现 / 传感器失效 / 多车混行）、自定义场景；
    另有 4.5 节「实车回放」类型。编码值映射见 `x-hunter-scene-types`。

    **场景配置结构**（4.2.2 节，不可更改）：场景元信息（scene_id / scene_name / scene_type /
    description / version / creator / tags）+ map / ego_vehicle / weather / actors[] / events[] /
    success_criteria / duration；字段与落库位置映射见 `x-hunter-scene-config-contract`。

    **状态机**：`draft → published → archived`（`scenes.status`，受控词表）；仅 `draft` 可编辑。

    **鉴权**：请求经网关校验 JWT 后转发，服务侧按 RBAC 资源域 `scene` 校验
    （permissions.resource=scene，action ∈ create/read/update/delete/execute）。

    **限流**：本服务无专属限流（附录 D 仅 3 条接口级限流，不含 scene），继承网关五级限流
    （全局 10000 QPS / 单用户 100 QPS / 单 IP 200 QPS）。

    **性能**：API 响应 P95 ≤ 200ms（13.2 节）；场景列表默认分页 20 条、上限 200 条。
  license:
    name: Proprietary - HunterEdge
servers:
  - url: https://hunter-edge.example.com
    description: 生产环境（经网关 /api/v1/scene 转发）
  - url: http://localhost:8081
    description: 本地开发（直连服务）
...
security:
  - bearerAuth: []
paths:
  /api/v1/scene:
    get:
      tags: [scenes]
      summary: 查询场景列表（分页 + 分类/状态/标签/关键字筛选）
      description: |
        场景库列表。默认按 `create_time DESC` 排序（命中索引
        `idx_scenes_status_type_create_time`）；`tags` 为包含查询（PostgreSQL `@>`，命中 GIN 索引）；
        `keyword` 对 `scene_name`/`description` 做不区分大小写的模糊匹配。
        软删除场景（`deleted_at` 非 NULL）一律不出现在结果中。
      operationId: listScenes
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: scene_type
          in: query
          required: false
          description: 场景类型（取值见 `x-hunter-scene-types`）
          schema: { $ref: "#/components/schemas/SceneType" }
        - name: status
          in: query
          required: false
          description: 场景状态（受控词表 draft/published/archived）
          schema: { $ref: "#/components/schemas/SceneStatus" }
        - name: tags
          in: query
          required: false
          description: 标签（可重复传参，多标签为 AND 关系）
          style: form
          explode: true
          schema:
            type: array
            maxItems: 20
            items: { type: string, minLength: 1, maxLength: 32 }
        - name: keyword
          in: query
          required: false
          description: 名称/描述模糊搜索关键字
          schema: { type: string, minLength: 1, maxLength: 64 }
        - name: creator
          in: query
          required: false
          description: 创建者 user_id（数据权限隔离时由服务端强制注入）
          schema: { type: string, format: uuid }
        - name: sort
          in: query
          required: false
          description: 排序字段（白名单，禁止任意字段排序以防注入）
          schema:
            type: string
            enum: [create_time, update_time, scene_name, version]
            default: create_time
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 场景列表
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneListResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      tags: [scenes]
      summary: 创建场景（场景编辑器保存）
      description: |
        创建场景，`status` 固定为 `draft`（不可由请求指定），`creator` 取自网关注入的
        `X-User-Id`，`version` 默认 `1.0.0`。
        `config` 结构必须与 4.2.2 节完全一致（见 `x-hunter-scene-config-contract`）。
        `scene_name` 在同一存活场景内唯一（`uq_scenes_scene_name`），冲突返回 3002。
      operationId: createScene
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SceneCreateRequest" }
      responses:
        "201":
          description: 创建成功（返回完整场景对象，status=draft）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "409":
          description: 场景名称已存在（code=3002）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scene/templates:
    get:
      tags: [templates]
      summary: 查询场景模板列表
      description: |
        返回 4.2.1 分类体系下的预置场景模板（基础 / 交互 / 环境 / 极端场景），用于场景编辑器一键创建；
        模板 `config` 为 4.2.2 结构，可直接作为 `POST /api/v1/scene` 的 `config` 提交。
        `real_vehicle`（实车回放）为 4.5 节自动提取产物，无预置模板，故不在 `category` 枚举中。
        ⚠ 模板数据来源（内置静态清单 / 数据库 / MinIO 资源）设计文档未定义，
        见 `x-hunter-pending-confirmation`。
      operationId: listSceneTemplates
      parameters:
        - name: category
          in: query
          required: false
          description: 模板分类（4.2.1 节分类）
          schema:
            type: string
            enum: [basic, interactive, environment, corner_case, custom]
        - name: scene_type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/SceneType" }
      responses:
        "200":
          description: 模板列表（不分页：模板数量受 4.2.1 分类体系约束）
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SceneTemplateListResponse" }
        "422": { $ref: "#/components/responses/InvalidParameter" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token（2 小时，网关签发）。载荷：`sub`(user_id) / `username` / `roles` / `permissions` /
        `iat` / `exp` / `iss` / `jti`。网关校验后注入 `X-User-Id`、`X-Roles`、`X-Trace-Id`，
        服务侧据此执行 RBAC（`permissions.resource=scene`，action ∈ create/read/update/delete/execute）
        与数据权限（按 creator/项目隔离）。
    deviceCertificate:
      type: mutualTLS
      description: |
        车辆设备认证：X.509 双向 TLS，证书 CommonName = vehicle_id。
        本服务无车端 REST 端点，此处仅声明全系统安全约定（车-云数据通道为 Kafka SASL_SSL）。
  parameters:
...
    PageQuery:
      name: page
      in: query
      required: false
      description: 页码（≥1）
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      description: 每页条数（1-200，与 hunter_common BaseRepository.paginate 上限一致）
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
  headers:
    RequestId:
      description: 链路追踪 ID（与请求头 `X-Request-ID` 一致）
      schema: { type: string, maxLength: 64 }
  responses:
    Unauthenticated:
      description: 未认证（code=1001，Token 缺失或无效）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 1001
            message: 未认证
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
...
    Forbidden:
      description: 无权限（code=1002，RBAC 资源域 scene 权限不足）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 1002
            message: 无权限
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
    InvalidParameter:
      description: 参数错误（code=2001，请求体/查询参数校验失败；HTTP 422，见 x-hunter-error-status-map）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 2001
            message: 参数错误
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
...
    ServiceUnavailable:
      description: 服务不可用（code=5001）：PostgreSQL / Redis / MinIO / Carla 管理 API 不可达
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 5001
            message: 服务不可用
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
    InternalError:
      description: 服务器内部错误（code=5000，未预期异常）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiResponse" }
          example:
            code: 5000
            message: 服务器内部错误
            data: null
            request_id: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
            timestamp: 1724035200
  schemas:
...
    ErrorCode:
      type: integer
      description: |
        预定义错误码（附录 A，禁止自定义）。本服务可能返回的子集：
        `0` 成功 / `1001` 未认证 / `1002` 无权限 / `1003` Token 过期 /
        `2001` 参数错误 / `2002` 参数缺失 / `3001` 资源不存在 / `3002` 资源已存在 /
        `3003` 资源状态冲突 / `5000` 服务器内部错误 / `5001` 服务不可用。
      enum: [0, 1001, 1002, 1003, 2001, 2002, 3001, 3002, 3003, 4001, 4002, 5000, 5001, 6001, 6002, 6003, 7001, 7002]
    ApiResponse:
      type: object
      description: 统一响应格式（所有业务响应/错误响应均为该结构）
      required: [code, message, data, request_id, timestamp]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          example: success
        data:
          nullable: true
          description: 业务数据；错误时为 null 或错误上下文
        request_id:
          type: string
          description: 请求追踪 ID（等于响应头 X-Request-ID）
          example: 8f9c2b7a-1d3e-4f5a-9b6c-0d1e2f3a4b5c
        timestamp:
          type: integer
          format: int64
          description: Unix epoch 秒
          example: 1724035200
...
    SceneStatus:
      type: string
      description: |
        场景状态受控词表（contracts/database/enums.md 第 6 节，DB CHECK 约束同源）。
        `draft` 可编辑；`published` 可下发/导出；`archived` 只读可删。
      enum: [draft, published, archived]
    SceneType:
      type: string
      description: |
        场景类型（4.2.1 分类体系的叶子场景；"分类 → 类型"映射见 `x-hunter-scene-types`）。
        ⚠ 编码值由本契约提出（设计文档未给出编码），需人工确认后落库，禁止前端硬编码分支。
      enum:
        - straight_cruise       # 基础场景：直线巡航
        - curve_driving         # 基础场景：弯道行驶
        - car_following         # 基础场景：跟车行驶
        - lane_change           # 基础场景：变道行驶
        - cut_in                # 交互场景：前车切入
        - pedestrian_crossing   # 交互场景：行人横穿
        - intersection_meeting  # 交互场景：路口会车
        - construction_detour   # 交互场景：施工绕行
        - rainy                 # 环境场景：雨天
        - night                 # 环境场景：夜间
        - foggy                 # 环境场景：雾天
        - backlight             # 环境场景：强光逆光
        - emergency_braking     # 极端场景：紧急制动
        - obstacle_appearance   # 极端场景：障碍物突现
        - sensor_failure        # 极端场景：传感器失效
        - multi_vehicle_mixed   # 极端场景：多车混行
        - custom                # 自定义场景
        - real_vehicle_replay   # 实车回放（4.5 节实车场景自动提取产物）
...
    SpawnPoint:
      type: object
      description: |
        生成点（自车/参与者）。⚠ 4.2.2 节仅给出 `spawn_point` 字段名，未定义内部结构；
        本契约按 Carla transform 语义定义（位置 m + 朝向 rad），需人工确认。
      required: [x, y, yaw]
      properties:
        x: { type: number, description: "X 坐标（m）" }
        y: { type: number, description: "Y 坐标（m）" }
        z: { type: number, default: 0.0, description: "Z 坐标（m）" }
        yaw: { type: number, minimum: -3.141592653589793, maximum: 3.141592653589793, description: "朝向（rad）" }
        lane_id: { type: string, nullable: true, description: "车道 ID（⚠ 语义待确认，可选）" }
    MapType:
      type: string
      description: "地图类型（⚠ 编码值为契约提出，设计文档未给出取值域）"
      enum: [carla_town, hd_map, site_map]
    SceneMap:
      type: object
      description: 场景地图段（4.2.2 节 map）
      required: [map_id, map_type, spawn_point]
      properties:
        map_id: { type: string, minLength: 1, maxLength: 128, description: "地图标识（Carla 地图名 / 高精地图 ID）" }
        map_type: { $ref: "#/components/schemas/MapType" }
        spawn_point: { $ref: "#/components/schemas/SpawnPoint" }
    EgoVehicle:
      type: object
      description: 自车初始状态段（4.2.2 节 ego_vehicle）
      required: [model, initial_speed, initial_steer]
      properties:
        model: { type: string, minLength: 1, maxLength: 64, example: HUNTER_SE, description: "车型（车端基线：HUNTER SE）" }
        initial_speed: { type: number, minimum: 0, description: "初始速度（m/s）" }
        initial_steer: { type: number, minimum: -1, maximum: 1, description: "初始转向（归一化 -1..1）" }
    Weather:
      type: object
      description: |
        环境条件段（4.2.2 节 weather，7 个字段全部必填）。
        取值范围按 Carla WeatherParameters 语义定义（设计文档未给出取值范围，见 x-hunter-pending-confirmation）。
      required: [cloudiness, rain, wetness, fog, wind, sun_azimuth, sun_altitude]
      properties:
        cloudiness: { type: number, minimum: 0, maximum: 100, description: "云量 %" }
        rain: { type: number, minimum: 0, maximum: 100, description: "降雨强度 %" }
        wetness: { type: number, minimum: 0, maximum: 100, description: "地面积水 %" }
        fog: { type: number, minimum: 0, maximum: 100, description: "雾浓度 %" }
        wind: { type: number, minimum: 0, maximum: 100, description: "风速强度 %" }
        sun_azimuth: { type: number, minimum: 0, maximum: 360, description: "太阳方位角（度）" }
        sun_altitude: { type: number, minimum: -90, maximum: 90, description: "太阳高度角（度，负值为夜间）" }
    ActorType:
      type: string
      description: "参与者类型（与遥测 perception.object_types 域一致：vehicle/pedestrian/other）"
      enum: [vehicle, pedestrian, other]
    ActorBehavior:
      type: object
      description: |
        参与者行为编排（4.2.2 节 actors[].behavior）。
        ⚠ behavior_type 取值域设计文档未定义（建议：lane_follow / speed_change / cut_in /
        lane_change / stop）；parameters 为行为参数自由结构；均需人工确认。
      required: [behavior_type]
      properties:
        behavior_type: { type: string, minLength: 1, maxLength: 64, example: lane_follow }
        parameters:
          type: object
          additionalProperties: true
          description: 行为参数（如目标速度、加减速度、触发距离），结构随 behavior_type 而定
    Actor:
      type: object
      description: 场景参与者（4.2.2 节 actors[]）
      required: [actor_id, type, spawn_point, behavior]
      properties:
        actor_id: { type: string, minLength: 1, maxLength: 64, description: "场景内参与者唯一标识" }
        type: { $ref: "#/components/schemas/ActorType" }
        spawn_point: { $ref: "#/components/schemas/SpawnPoint" }
        behavior: { $ref: "#/components/schemas/ActorBehavior" }
    SceneEventType:
      type: string
      description: |
        场景事件类型（4.2.2 节 events[].type）。
        ⚠ 设计文档未定义取值域；本契约按 4.2.1 场景与 4.5 实车触发事件提出候选值，需人工确认。
      enum:
        - collision
        - cut_in
        - pedestrian_crossing
        - emergency_brake
        - manual_takeover
        - sensor_failure
        - obstacle_appearance
    SceneEventTrigger:
      type: object
      description: |
        事件触发条件（4.2.2 节 events[].trigger）。
        ⚠ 内部结构设计文档未定义：condition 为触发表达式标识（如 OpenSCENARIO 条件名或
        `ego.position.x > 10`），parameters 为其参数。
      required: [condition]
      properties:
        condition: { type: string, minLength: 1, maxLength: 256 }
        parameters: { type: object, additionalProperties: true }
    SceneEventAction:
      type: object
      description: |
        事件触发后的动作（4.2.2 节 events[].action）。
        ⚠ 内部结构设计文档未定义：action_type 建议取 speed_change / lane_change / stop /
        emergency_stop / actor_spawn。
      required: [action_type]
      properties:
        action_type: { type: string, minLength: 1, maxLength: 64, example: speed_change }
        parameters: { type: object, additionalProperties: true }
    SceneEvent:
      type: object
      description: 场景事件（4.2.2 节 events[]）
      required: [event_id, type, trigger, action]
      properties:
        event_id: { type: string, minLength: 1, maxLength: 64, description: "场景内事件唯一标识" }
        type: { $ref: "#/components/schemas/SceneEventType" }
        trigger: { $ref: "#/components/schemas/SceneEventTrigger" }
        action: { $ref: "#/components/schemas/SceneEventAction" }
    SuccessCriteria:
      type: object
      description: 场景成功判据（4.2.2 节 success_criteria）
      required: [max_speed_deviation, no_collision, min_safe_distance]
      properties:
        max_speed_deviation: { type: number, minimum: 0, description: "最大速度偏差（m/s）" }
        no_collision: { type: boolean, default: true, description: "是否要求全程无碰撞（true=通过判据）" }
        min_safe_distance: { type: number, minimum: 0, description: "最小安全距离（m）" }
    SceneConfig:
      type: object
      description: |
        场景仿真配置（4.2.2 节结构中的仿真相关段，落库 scenes.config_json）。
        7 个字段全部必填，与 x-hunter-scene-config-contract 声明一致。
      required: [map, ego_vehicle, weather, actors, events, success_criteria, duration]
      properties:
        map: { $ref: "#/components/schemas/SceneMap" }
        ego_vehicle: { $ref: "#/components/schemas/EgoVehicle" }
        weather: { $ref: "#/components/schemas/Weather" }
        actors:
          type: array
          maxItems: 100
          description: 参与者列表（可为空数组）
          items: { $ref: "#/components/schemas/Actor" }
        events:
          type: array
          maxItems: 100
          description: 事件列表（可为空数组）
          items: { $ref: "#/components/schemas/SceneEvent" }
        success_criteria: { $ref: "#/components/schemas/SuccessCriteria" }
        duration: { type: number, exclusiveMinimum: 0, description: "场景时长（s）；⚠ 上限未定义，下发前校验" }
    SceneMeta:
      type: object
      description: |
        场景元信息（4.2.2 节结构中的元信息段，一一对应 scene_svc.scenes 列）。
        与 SceneConfig 合并即为 4.2.2 节完整场景配置结构。
      required: [scene_id, scene_name, scene_type, description, version, creator, tags, status, create_time, update_time]
      properties:
        scene_id: { type: string, format: uuid }
        scene_name: { type: string, minLength: 1, maxLength: 128 }
        scene_type: { $ref: "#/components/schemas/SceneType" }
        description: { type: string, nullable: true, maxLength: 1024 }
        version: { type: string, pattern: "^\\d+\\.\\d+\\.\\d+$", example: 1.0.0, description: "语义化版本（导出时写入 x-scenario-version）" }
        creator: { type: string, format: uuid, description: "创建者 user_id（逻辑外键 → user_svc.users）" }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 32 }
        status: { $ref: "#/components/schemas/SceneStatus" }
        create_time: { type: string, format: date-time }
        update_time: { type: string, format: date-time }
    Scene:
      description: 场景完整对象 = SceneMeta（元信息，4.2.2 元信息段）+ config（仿真配置段）
      allOf:
        - $ref: "#/components/schemas/SceneMeta"
        - type: object
          required: [config]
          properties:
            config: { $ref: "#/components/schemas/SceneConfig" }
    SceneTemplate:
      type: object
      description: 场景模板（4.2.1 分类体系预置；config 可直接作为创建请求的 config 提交）
      required: [template_id, template_name, scene_type, description, config]
      properties:
        template_id: { type: string, minLength: 1, maxLength: 64 }
        template_name: { type: string, minLength: 1, maxLength: 128 }
        scene_type: { $ref: "#/components/schemas/SceneType" }
        description: { type: string, nullable: true }
        config: { $ref: "#/components/schemas/SceneConfig" }

    # ---------- 业务数据段（统一响应 data 字段） ----------
    SceneListData:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Scene" }
        total: { type: integer, minimum: 0, description: "满足筛选条件的总数（软删除数据不计入）" }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 200 }
    SceneTemplateListData:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SceneTemplate" }
        total: { type: integer, minimum: 0 }
...
    SceneResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/Scene" }
    SceneListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneListData" }
    SceneTemplateListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/SceneTemplateListData" }
...
    SceneCreateRequest:
      type: object
      description: |
        创建场景请求：元信息（写入 scenes 列）+ config（写入 config_json，4.2.2 结构）。
        status / creator / version / create_time 由服务端生成，不接收客户端传值。
      required: [scene_name, scene_type, config]
      additionalProperties: false
      properties:
        scene_name: { type: string, minLength: 1, maxLength: 128 }
        scene_type: { $ref: "#/components/schemas/SceneType" }
        description: { type: string, nullable: true, maxLength: 1024 }
        tags:
          type: array
          maxItems: 20
          default: []
          items: { type: string, minLength: 1, maxLength: 32 }
        config: { $ref: "#/components/schemas/SceneConfig" }
...